+++ src/cli/src/modules/refactor/index.ts
@@ -1,4 +1,5 @@
 export * from "./bridge-factory.js";
 export * from "./parser-bridge.js";
+export * from "./project-metadata-mutation.js";
 export * from "./semantic-bridge.js";
 export * from "./transpiler-bridge.js";
